Logistic Regression/requirements.txt
numpy~=1.19.3
matplotlib~=3.3.3
Pillow~=8.0.1
h5py~=3.1.0